refactor: change some files place
