--- react-components/src/api/axios.ts
@@ -1,7 +1,0 @@
-import axios from 'axios';
-
-const axiosInstance = axios.create({
-  baseURL: 'https://api.unsplash.com',
-});
-
-export default axiosInstance;
